PID worky sorta need odom pod but you know in theory it all good hopefully that true and we don't end up dying woohoo, did you even read this far? lol :)

+++ include/SSOV/controller/BoomerangController.hpp
@@ -30,7 +30,7 @@ namespace ssov {
                 return std::make_shared<BoomerangController>(linear_constants, angular_constants, min_error, lead_pct);
             }
 
-            DriveSignal compute(const Pose &current_pose, const Pose &target_pose, bool reverse, bool thru, bool holonomic, int strafe_angle=0) override;
+            DriveSignal compute(const Pose &current_pose, const Pose &target_pose, bool reverse, bool thru, bool holonomic, float strafe_angle=0) override;
             void reset() override;
             void set_debug(bool debug) {
                 this->debug = debug;
